Split Populations into separate files
