+++ evol/population/__init__.py
@@ -1,0 +1,2 @@
+from .base import Population
+from .contest import ContestPopulation
